Fix rebalancing leading to consumer not working.
If a rebalance happens while a poll is taking place, a new
RequestMessages might never be sent and if there is no
overlap between the new set of assigned partitions and the
old set of partitions, the consumer might be effectively dead.
The KafkaConsumerActor will keep polling on the old partitions
and never send anything back that would trigger a new
RequestMessages. Setting requested to false when partitions are
assigned fixes the problem.