Add a test for rebalancing during requests
The test starts up multiple consumers listening to a topic with
multiple partitions and no messages to begin with. After the initial
assignment, another consumer is added, rebalance happens and a
unique message is sent to each partition. If all unique messages are
received, the test succeeds. Removing 'requests = false' from the
revoking and assigning of partitions methods in SingleSourceLogic
causes this test to fail.