Fix mailbox flooding with Poll messages in KafkaConsumerActor (#236)
If the execution time of `poll()` is longer than the configured
`pollInterval`, the actor's mailbox could be flooded with `Poll`
messages, starving out other messages.

This commit changes the scheduling mechanism to use `scheduleOnce`: a
`Poll` message will be scheduled in `preStart()`, and another will be
scheduled once the `Poll` message was processed.