Discard unsent ProducerRecords after encountering a send-failure in ProducerStage #295
* The `Supervision.Stop`-behaviour of the `ProducerStage` flushed all unsent `ProducerRecord`s
queued in the `KafkaProducer`s send-buffer when encountering a send-failure, regardless of their
chronological order in relation to the failed `ProducerRecord`.
As the stream was shutdown after encountering the failure, these `ProducerRecord`s were produced
to Kafka but could never be observed on stages downstream.
This situation lead to a non-deterministic failure-behaviour where a client could not be sure which
`ProducerRecord`s were actually produced to Kafka after encountering a send-failure.
This behaviour also lead to message-reordering in regards to the message-source, as failing
`ProducerRecord`s were dropped but subsequent `ProducerRecord`s were produced to Kafka.

* Introduce deterministic failure-behaviour by closing the `KafkaProducer` in the send-callback
upon encountering a send-failure. This behaviour ensures that no `ProducerRecord` is sent to Kafka
which was scheduled after the failed `ProducerRecord`.

* Apply the new failure-behaviour only if the `ProducerStage` is stopped by the send-failure,
determined by the given `SupervisionStrategy`.

Refs #295

Lightbend-CLA has been signed.