Additional logging around reconciliation and wake up expcetions
Various users are seeing WakeUpExceptions even though it is only
triggered after 3 seconds. This appears to be an issue with the
Kafka client library.

This adds a thread dump and more detailed logging around partition
assignment to better able us to raise issues against the kafka
client.

The thread dump log may be too verbose so can be turned off in config.