Use the simple shutdown mechanism for external offset storage and the complex mechanism for kafka offset storage
The complex shutdown is necessary regardless of batching to ensure that the Consumer actor is not shutdown before the stream has finished processing the last message from the `Source` and produced a commit request for it.

Even when not using batching, when there are asynchronous boundaries between the `Source` and the call to `commitScaladsl`, there is a race condition between shutting down the consumer actor and producing the commit request. To keep the documentation simpler, I recommend to always use the complex shutdown mechanism when using offset storage in Kafka.